+++ Packages/ai.fxn.fxn3d/Plugins/iOS/Function.framework/Headers/FXNVersion.h
@@ -12,7 +12,7 @@
 
 #define FXN_VERSION_MAJOR 0
 #define FXN_VERSION_MINOR 0
-#define FXN_VERSION_PATCH 25
+#define FXN_VERSION_PATCH 30
 
 /*!
  @function FXNGetVersion